requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic>=2.9.2
python-dotenv==1.0.1
tomli==2.0.1
requests==2.32.3
jinja2==3.1.4
faiss-cpu==1.8.0.post1
numpy==1.26.4
pandas==2.2.2
sentence-transformers==2.7.0
transformers==4.42.4
peft==0.11.1
accelerate==0.33.0
datasets==2.20.0
torch>=2.1.0